Change usize to isize

+++ examples/stdout.rs
@@ -1,11 +1,11 @@
 use sys_call::sys_call;
 
 fn main() {
-    const WRITE: usize = 1;
-    const STDOUT: usize = 1;
+    const WRITE: isize = 1;
+    const STDOUT: isize = 1;
     let buf = b"Hello World!\n";
 
     unsafe {
-        sys_call!(WRITE, STDOUT, buf.as_ptr() as usize, buf.len());
+        sys_call!(WRITE, STDOUT, buf.as_ptr() as isize, buf.len() as isize);
     }
 }
